Recovered from Eclipse.
git-svn-id: https://subversion.assembla.com/svn/cmosher@4655 884d43a4-ffbe-47e9-90ca-fd8c37d7eb32

+++ GDiffView/src/nu/mine/mosher/GDiffCopy.java
@@ -17,7 +17,7 @@ public class GDiffCopy extends GDiffCmd
     public GDiffCopy(long pos, long end)
     {
         this.pos = pos;
-        this.len = end-pos;
+        this.len = (int)(end-pos);
     }
 	public long getPosition()
 	{
